Switch to using project-specific webpack config instead of global config
This isn't ideal, but running Webpack was causing the "TypeError: The 'compilation' argument must be an instance of Compilation" error that's typically caused by two different versions of Webpack running. I only found one version of Webpack installed, and no other solution worked except for this. Might be worth revisiting later and overhauling the build process to allow reuse of global Webpack config files.

The image upload project will likely need this same treatment but is currently left as-is.